src/aws-lambda/roomgenerator/src/_layers/shared/requirements.txt
opensearch-py==2.4.2
pillow==10.2.0
aws-lambda-powertools==2.33.1
boto3==1.34.44
botocore==1.34.44
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
pydantic==2.6.2
pydantic_core==2.16.3
crhelper